+++ src/test/scala/com/atomist/rug/runtime/plans/AnnotatedRugFunctionTest.scala
@@ -5,6 +5,7 @@ import com.atomist.rug.spi.Handlers.Status
 import org.scalatest.{FlatSpec, Matchers}
 
 class AnnotatedRugFunctionTest extends FlatSpec with Matchers{
+
   it should "report the right field values in an annotation function" in {
     val fn = new ExampleAnnotatedRugFunction()
     assert(fn.name === ExampleAnnotatedRugFunction.theName)
@@ -14,7 +15,7 @@ class AnnotatedRugFunctionTest extends FlatSpec with Matchers{
     assert(fn.secrets.size === 1)
     assert(fn.secrets.head.name === "user_token")
     assert(fn.secrets.head.path === ExampleAnnotatedRugFunction.theSecretPath)
-    assert(fn.parameters.size === 2)
+    assert(fn.parameters.size ===4)
     assert(fn.parameters.head.name === "number")
 
     val res = fn.run(SimpleParameterValues(SimpleParameterValue("number", "100"), SimpleParameterValue("user_token", "woot")))
@@ -22,5 +23,26 @@ class AnnotatedRugFunctionTest extends FlatSpec with Matchers{
 
     val failed = fn.run(SimpleParameterValues(SimpleParameterValue("number", 100.asInstanceOf[AnyRef]), SimpleParameterValue("user_token", "foo")))
     assert(failed.status === Status.Failure)
+  }
+
+  it should "optional string parameters with defaults should not end up as null" in {
+    val fn = new ExampleAnnotatedRugFunction()
+    val res = fn.run(SimpleParameterValues(SimpleParameterValue("number", "100"), SimpleParameterValue("user_token", "woot")))
+    assert(res.status === Status.Success)
+    assert(fn.blah == "")
+  }
+
+  it should "optional int parameters with no default gets converted to 0" in {
+    val fn = new ExampleAnnotatedRugFunction()
+    val res = fn.run(SimpleParameterValues(SimpleParameterValue("number", "100"), SimpleParameterValue("user_token", "woot")))
+    assert(res.status === Status.Success)
+    assert(fn.other == 0)
+  }
+
+  it should "optional int parameters with default gets converted properly" in {
+    val fn = new ExampleAnnotatedRugFunction()
+    val res = fn.run(SimpleParameterValues(SimpleParameterValue("number", "100"), SimpleParameterValue("user_token", "woot")))
+    assert(res.status === Status.Success)
+    assert(fn.yum == 42)
   }
 }