+++ src/test/scala/com/atomist/rug/runtime/plans/ExampleAnnotatedRugFunction.scala
@@ -10,10 +10,18 @@ import com.atomist.rug.spi.{AnnotatedRugFunction, FunctionResponse}
 class ExampleAnnotatedRugFunction
   extends AnnotatedRugFunction {
 
+  var blah: String = _
+  var other: Int = -1
+  var yum: Int = -1
   @RugFunction(name = "example-function", description = "Description of function", tags = Array(new Tag(name = "tag content")))
   def invoke(@Parameter (name="number") number: Int,
              @Parameter (name="blah", required = false) blah: String,
+             @Parameter (name="other", required = false) other: Int,
+             @Parameter (name="yum", required = false, defaultValue = "42") yum: Int,
              @Secret(name = "user_token", path = "github/user_token=repo") user_token: String): FunctionResponse = {
+    this.blah = blah
+    this.other = other
+    this.yum = yum
     if(number == 100 && user_token == "woot"){
       FunctionResponse(Status.Success)
     }else{