+++ CHANGELOG.md
@@ -10,8 +10,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 [Unreleased]: https://github.com/atomist/rug/compare/1.0.0-m.5...HEAD
 
 ### Added 
+
 -   Added `GITHUB_DEFAULT_REPO_VISIBILITY` mapped parameter that determines
     repository visibility based on organization GitHub plan
+
+### Fixed
+
+-   Ensure default values for optional parameters to Rug Functions
+    are used, instead of null. Primitives are converted to 0 or false
 
 ## [1.0.0-m.5] - 2017-07-11